WIP: Refactor start from OXC Visitor to OXC Traverse.
