+++ Cargo.toml
@@ -4,8 +4,15 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-oxc = { version = "*", features = ["full"] }
+#oxc = { version = "*", features = ["full"] }
 oxc_index = "2.0.0"
+oxc_parser = "0.52.0"
+oxc_ast = "0.52.0"
+oxc_codegen = "0.52.0"
+oxc_allocator = "0.52.0"
+oxc_semantic = "0.52.0"
+oxc_span = "0.52.0"
+oxc_traverse = "0.52.0"
 thiserror = "2.0.11"
 base64 = "0.22.1"
 markup5ever = "0.14.1"
