+++ src/component/mod.rs
@@ -1,0 +1,12 @@
+mod common;
+mod qwik_component;
+mod qrl;
+mod id;
+mod source_info;
+
+// -- Flatten
+pub(crate) use common::*;
+pub(crate) use qwik_component::*;
+pub(crate) use qrl::*;
+pub(crate) use id::*;
+pub(crate) use source_info::*;
