+++ src/error.rs
@@ -13,5 +13,5 @@ pub enum Error {
 
 
     #[error(transparent)]
-    OxcUnknownExtension(#[from] oxc::span::UnknownExtension),
+    OxcUnknownExtension(#[from] oxc_span::UnknownExtension),
 }
