+++ src/lib.rs
@@ -1,22 +1,8 @@
-mod prelude;
-mod error;
-mod parser;
-pub(crate) mod component;
-mod transformer;
 mod sources;
-mod ast_builder_ext;
+pub mod component;
+pub(crate)  mod error;
+pub(crate) mod ext;
+pub(crate) mod prelude;
 
-pub fn add(left: u64, right: u64) -> u64 {
-    left + right
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn it_works() {
-        let result = add(2, 2);
-        assert_eq!(result, 4);
-    }
-}
+pub mod transformer;
+mod traverser;
